pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "namesbench"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "langchain",
  "langchain-openai",
  "langchain-anthropic",
  "langchain-google-genai",
  "pydantic",
  "python-dotenv",
  "Pillow",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
run-benchmark = "namesbench.run_benchmark:main"

